test(service): inline category mock repository into category_service_test
